backend/requirements-minimal.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
python-multipart==0.0.20
python-dotenv==1.1.1
httpx==0.28.1
requests==2.32.4
PyYAML==6.0.2
# Core (light) dependencies – heavy ML libs moved to requirements-ml.txt (optional)
numpy==2.3.1
# For PII detection
presidio-analyzer==2.2.359
presidio-anonymizer==2.2.359
spacy==3.8.7
# NOTE: spaCy language models (fr_core_news_sm, en_core_web_sm) are downloaded at runtime if missing.
# Essential utilities
click==8.2.1
colorama==0.4.6
pymysql==1.1.1
openai==1.35.10
tiktoken==0.7.0
